+++ components/FavoriteButton/ButtonsStyle.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+// Zur Anordnung der Buttons überhalb des Film Posters
+export const ButtonFav = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-bottom: 5px;
+`;
