--- components/FavoriteButton/FavoriteButton.css
@@ -1,23 +1,0 @@
-.favorite-button {
-  position: absolute;
-  top: 0;
-  right: 0;
-  appearance: none;
-  background: none;
-  border: none;
-  padding: 0;
-  border-radius: 999px;
-  aspect-ratio: 1;
-  transition: background-color 0.2s ease-in-out;
-  background-color: transparent;
-  margin: -6px;
-  padding: 6px;
-}
-
-.favorite-button:hover {
-  background-color: var(--color-water);
-}
-
-.favorite-button:active {
-  background-color: var(--color-water-10);
-}
